Organized code so that we have a geostatsmodels namespace for imports, added setup script, and explicit MIT license file
