+++ geostatsmodels/simulation.py
