+++ geostatsmodels/utilities.py
